requirements.txt
langchain_community
openai
langgraph
python-dotenv
langchain-cerebras
langchain_huggingface
langchain
bs4
langchain_openai
guardrails-ai
smolagents
wikipedia
google-genai
pypdf
unstructured
faiss-cpu
gradio
sentence-transformers